handling error feedback more gracefully
