--- source/resources/generalScreens/errorScreen.ts
@@ -1,38 +1,0 @@
-import { errorColor, secondaryColor } from "@utilities/ConsoleService.js";
-
-export function getErrorScreen() {
-  const e = (text: string) => errorColor(text);
-  const s = (text: string) => secondaryColor(text);
-  return s(`*******************************************************************************
-          |                   |                  |             \\_______/
- _________|___________________|__________________|_________\`.,-'\\_____/\`-.,'___
-|                   |                   |                  /  /\`.,' \`.,'\\  \\
-|___________________|___________________|_________________/__/__/     \\__\\__\\__
-          |                   |                  |        \\  \\  \\     /  /  /
- _________|___________________|__________________|_________\\  \\,'\`._,'\`./  /___
-|                   |  ${e(
-    "/‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾\\"
-  )}     \\,'\`./___\\,'\`./
-|___________________|__${e(
-    "⎸   _____                      ⎹"
-  )}____,'\`-./_____\\,-'\`.___
-          |            ${e(
-            "⎸  | ____|_ __ _ __ ___  _ __  ⎹"
-          )}        /       \\
- _________|____________${e(
-   "⎸  |  _| | '__| '__/ _ \\| '__| ⎹"
- )}________________|_______
-|                   |  ${e("⎸  | |___| |  | | | (_) | |    ⎹")}    |
-|___________________|__${e(
-    "⎸  |_____|_|  |_|  \\___/|_|    ⎹"
-  )}____|___________________
-____/______/______/____${e(
-    "\\______________________________/"
-  )}_____/______/______/____
-/______/______/______/______/______/______/______/______/______/______/______/_
-____/______/______/______/______/______/______/______/______/______/______/____
-/______/______/______/______/______/______/______/______/______/______/______/_
-____/______/______/______/______/______/______/______/______/______/______/____
-/______/______/______/______/______/______/______/______/______/______/______/_
-*******************************************************************************`);
-}
