+++ source/resources/templates/templates.ts
@@ -1,8 +1,8 @@
-import { primaryColor, secondaryColor } from "@utilities/ConsoleService.js";
+import { errorColor, primaryColor, secondaryColor } from "@utilities/ConsoleService.js";
 import { getTerm } from "@utilities/LanguageService.js";
 import packageJson from "@root/package.json" with {type: "json"};
 
-export default function getTemplateRoomAscii() {
+function getTemplateRoomAscii() {
   const p = (text: string) => primaryColor(text);
   const s = (text: string) => secondaryColor(text);
   return s(`*******************************************************************************
@@ -28,7 +28,7 @@ ____/______/______/______/______/______/______/______/______/______/______/____
 }
 
 
-export const getErrorAscii = () => `/========================================================\\
+const getErrorAscii = () => `/========================================================\\
 ||                                                      ||
 ||    ______    ______    ______    ______    ______    ||
 ||   /      \\  /      \\  /      \\  /      \\  /      \\   ||
@@ -40,7 +40,7 @@ export const getErrorAscii = () => `/===========================================
 ||                                                      ||
 \\========================================================/`;
 
-export const getWarningAscii = () => `/==================================================\\
+const getWarningAscii = () => `/==================================================\\
 ||                                                ||
 ||   __   __   __   ______    ______   _______    ||
 ||  /  | /  | /  | /      \\  /      \\ /       \\   ||
@@ -52,7 +52,7 @@ export const getWarningAscii = () => `/=========================================
 ||                                                ||
 \\==================================================/`;
 
-export const getTitle = () => 
+const getTitle = () => 
 ` /$$$$$$$            /$$$$$$$           /$$$$$$  /$$       /$$$$$$
 | $$__  $$          | $$__  $$         /$$__  $$| $$      |_  $$_/
 | $$  \\ $$ /$$$$$$$ | $$  \\ $$        | $$  \\__/| $$        | $$  
@@ -64,7 +64,7 @@ export const getTitle = () =>
 
 ${packageJson.version} - ${getTerm("welcome")}`;
 
-export const testCombatBackground = [
+const testCombatBackground = [
   "********************************************************************************",
   "                      /‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾&                      ",
   "                      | Goblin [|||||||||||||||||] 10/10 |                      ",
@@ -86,3 +86,79 @@ export const testCombatBackground = [
   "/  ' .        _.=*                                            *=._        . '  &",
   "********************************************************************************"
 ].join("\n")
+
+function getErrorScreen() {
+  const p = (text: string) => primaryColor(text);
+  const s = (text: string) => secondaryColor(text);
+  return s(`*******************************************************************************
+          |                   |                  |             \\_______/
+ _________|___________________|__________________|_________\`.,-'\\_____/\`-.,'___
+|                   |                   |                  /  /\`.,' \`.,'\\  \\
+|___________________|___________________|_________________/__/__/     \\__\\__\\__
+          |                   |                  |        \\  \\  \\     /  /  /
+ _________|___________________|__________________|_________\\  \\,'\`._,'\`./  /___
+|                   |  ${p(
+    "/‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾\\"
+  )}     \\,'\`./___\\,'\`./
+|___________________|__${p(
+    "⎸   _____                      ⎹"
+  )}____,'\`-./_____\\,-'\`.___
+          |            ${p(
+            "⎸  | ____|_ __ _ __ ___  _ __  ⎹"
+          )}        /       \\
+ _________|____________${p(
+   "⎸  |  _| | '__| '__/ _ \\| '__| ⎹"
+ )}________________|_______
+|                   |  ${p("⎸  | |___| |  | | | (_) | |    ⎹")}    |
+|___________________|__${p(
+    "⎸  |_____|_|  |_|  \\___/|_|    ⎹"
+  )}____|___________________
+____/______/______/____${p(
+    "\\______________________________/"
+  )}_____/______/______/____
+/______/______/______/______/______/______/______/______/______/______/______/_
+____/______/______/______/______/______/______/______/______/______/______/____
+/______/______/______/______/______/______/______/______/______/______/______/_
+____/______/______/______/______/______/______/______/______/______/______/____
+/______/______/______/______/______/______/______/______/______/______/______/_
+*******************************************************************************`);
+}
+
+function getWarningScreenAscii() {
+  const e = (text: string) => errorColor(text);
+  const s = (text: string) => secondaryColor(text);
+  return s(`*******************************************************************************
+          |                   |                  |             \\_______/       
+ _________|___________________|__________________|_________\`.,-'\\_____/\`-.,'___
+|                   |                   |                  /  /\`.,' \`.,'\\  \\   
+|___________________|___________________|_________________/__/__/     \\__\\__\\__
+          |                   |                  |        \\  \\  \\     /  /  /
+ _________|______${e(
+   "/‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾\\"
+ )}\\,'\`._,'\`./  /___
+|                ${e(
+    "⎸                            _              ⎹"
+  )}'\`./___\\,'\`./
+|________________${e(
+    "⎸  __      ____ _ _ __ _ __ (_)_ __   __ _  ⎹"
+  )}-./_____\\,-'\`.___
+          |      ${e(
+            "⎸  \\ \\ /\\ / / _` | '__| '_ \\| | '_ \\ / _` | ⎹"
+          )} /       \\
+ _________|______${e(
+   "⎸   \\ V  V / (_| | |  | | | | | | | | (_| | ⎹"
+ )}_________|_______
+|                ${e("⎸    \\_/\\_/ \\__,_|_|  |_| |_|_|_| |_|\\__, | ⎹")}
+|________________${e(
+    "⎸                                    |___/  ⎹"
+  )}_________________
+____/______/_____${e(
+    "\\___________________________________________/"
+  )}_____/______/____
+/______/______/______/______/______/______/______/______/______/______/______/_
+____/______/______/______/______/______/______/______/______/______/______/____
+/______/______/______/______/______/______/______/______/______/______/______/_
+____/______/______/______/______/______/______/______/______/______/______/____
+/______/______/______/______/______/______/______/______/______/______/______/_
+*******************************************************************************`);
+}
